@@ -5,10 +5,10 @@ apply plugin: 'realm-android'
 
 ext {
     androidTestVer = "0.5"
-    buildToolsVer = "26.0.2"
+    buildToolsVer = "27.0.1"
     butterknifeVer = "8.8.1"
     checkstyleVer = '7.7'
-    compileSdkVer = 25
+    compileSdkVer = 26
     constraintVer = "1.0.2"
     dexMakerVer = "1.2"
     espContVer = "2.2.2"
@@ -25,8 +25,8 @@ ext {
     mockitoVer = "2.9.0"
     roboElecVer = "3.4.2"
     rxJavaVer = "1.3.0"
-    supportVer = "25.4.0"
-    targetSdkVer = 25
+    supportVer = "27.0.0"
+    targetSdkVer = 26
     verCode = 1
     verName = "1.0"
 }
@@ -59,6 +59,9 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
             multiDexKeepFile file("${configDir}/multidex/multidex.txt")
         }
+    }
+    dexOptions {
+        preDexLibraries = false
     }
     packagingOptions {
         exclude 'META-INF/rxjava.properties'
@@ -100,29 +103,30 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    compile "com.android.support:appcompat-v7:${supportVer}"
-    compile "com.android.support:support-v4:${supportVer}"
-    compile "com.android.support:design:${supportVer}"
-    compile "com.android.support:customtabs:${supportVer}"
-    compile "com.android.support.constraint:constraint-layout:${constraintVer}"
-    compile "com.android.support:cardview-v7:${supportVer}"
+    implementation "com.android.support:appcompat-v7:${supportVer}"
+    implementation "com.android.support:support-v4:${supportVer}"
+    implementation "com.android.support:design:${supportVer}"
+    implementation "com.android.support:customtabs:${supportVer}"
+    implementation "com.android.support.constraint:constraint-layout:${constraintVer}"
+    implementation "com.android.support:cardview-v7:${supportVer}"
 
-    compile "com.jakewharton:butterknife:${butterknifeVer}"
+    implementation "com.jakewharton:butterknife:${butterknifeVer}"
     annotationProcessor "com.jakewharton:butterknife-compiler:${butterknifeVer}"
-    compile "com.google.firebase:firebase-core:${firebaseVer}"
-    compile "com.google.firebase:firebase-database:${firebaseVer}"
-    compile "com.google.firebase:firebase-storage:${firebaseVer}"
-    compile "com.google.firebase:firebase-auth:${firebaseVer}"
-    compile "com.google.android.gms:play-services-auth:${playServicesVer}"
-    compile "com.google.firebase:firebase-messaging:${firebaseVer}"
+    implementation "com.google.firebase:firebase-core:${firebaseVer}"
+    implementation "com.google.firebase:firebase-database:${firebaseVer}"
+    implementation "com.google.firebase:firebase-storage:${firebaseVer}"
+    implementation "com.google.firebase:firebase-auth:${firebaseVer}"
+    implementation "com.google.android.gms:play-services-auth:${playServicesVer}"
+    implementation "com.google.firebase:firebase-messaging:${firebaseVer}"
 
-    compile "com.github.bumptech.glide:glide:${glideVer}"
+    implementation "com.github.bumptech.glide:glide:${glideVer}"
     annotationProcessor "com.github.bumptech.glide:compiler:${glideVer}"
+    implementation 'com.lapism:searchview:5.0.0-alpha7'
+    implementation 'com.google.android:flexbox:0.3.1'
 
-    compile "io.reactivex.rxjava2:rxjava:2.1.6"
+    implementation "io.reactivex.rxjava2:rxjava:2.1.6"
 
     //Unit test compiles
     testCompile "junit:junit:${junitVer}"
